apps/agni-ai/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.129.0
fastapi-cli==0.0.21
fastapi-cloud-cli==0.12.0
fastar==0.8.0
fastembed==0.7.4
filelock==3.24.2
flatbuffers==25.12.19
fsspec==2026.2.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-core==1.2.13
langchain-ollama==1.0.1
langchain-openai==1.1.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.3
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
numpy==2.4.2
ollama==0.6.1
onnxruntime==1.24.1
openai==2.21.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.33.5
py_rust_stemmers==0.1.5
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.2
qdrant-client==1.16.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rich-toolkit==0.19.4
rignore==0.6.4
rpds-py==0.27.0
sentry-sdk==2.52.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.14.1
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==15.0.1
win32_setctime==1.2.0
xxhash==3.6.0
zstandard==0.25.0
